requirements.txt
flask
flask-restx
langchain
langchain-community
langchain-text-splitters
sentence-transformers
pinecone-client
langchain-groq
python-dotenv
langchain-huggingface